Rename Router -> routeRequest to better fit actual shape
