+++ src/main/routes/files/static-file-route.injectable.ts
@@ -2,7 +2,7 @@
  * Copyright (c) OpenLens Authors. All rights reserved.
  * Licensed under MIT License. See LICENSE in root directory for more information.
  */
-import { getRouteInjectable } from "../../router/router.injectable";
+import { getRouteInjectable } from "../../router/route-request.injectable";
 import isDevelopmentInjectable from "../../../common/vars/is-development.injectable";
 import { route } from "../../router/route";
 import prodStaticFileRouteHandlerInjectable from "./production.injectable";
